taha_back/requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-cors-headers==3.7.0
django-filter==2.4.0
django-graphql-jwt==0.3.1
django-treebeard==4.4
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
idna==2.10
ippanel==1.1.0
ordereddict==1.1
Pillow==8.1.0
promise==2.3
PyJWT==1.7.0
pytz==2021.1
psycopg2
requests==2.25.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.3
WooCommerce==2.1.1

